python-services/face_recognition/requirements.txt
# Face Recognition Service Dependencies
mediapipe==0.10.21
protobuf==4.25.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
aiofiles==23.2.1
